Allow which digraphs to replace
The idea is, to only replace those chars, that are specified and leave
all other characters alone for easier transformation of larger selected
texts.

Do this by specifying the global variable g:Unicode_ConvertDigraphSubset.
All digraphs, that are defined in that list, will be replaced, the resst
will be left alone (closes #8)