Introduce data and cache directory
Introduce a data and a cache directory. The UnicodeData.txt (as
downloaded from unicode.org) is stored in the data directory. The cache
directory contains the files which are derived from this source file,
currently these are UnicodeData.vim and UnicodeTable.txt.

When running Neovim the default for the cache and data directory is
following the XDG Base Directory Specification. In this case the
directories usually are ~/.local/share/nvim/site/unicode and
~/.cache/nvim/unicode.

When not running Neovim, the directory defaults are the old directory
"unicode" under the autoload directory where the plugin is installed.

Also allow the user to set the data and cache directories explicitly in
their config file via the two variables g:Unicode_data_directory and
g:Unicode_cache_directory.