Trim the user's number input.
When scrolling through a long list of `:UnicodeSearch!` results, I
sometimes forget to erase the extra spaces that end up in my answer to
the 'Enter number of char to insert:' prompt. Thus, my answer ends up
becoming something like '    42', which doesn't match the `^\d\+`
pattern, and is rejected. This commit trims the answer so that the match
is successful. There are two places where the input() function is
called; both were fixed.